+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+
+- Update example app to React Native 0.83
+
 ## [0.17.3] - 2026-01-03
 
 ## Fixed