+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update example app to React Native 0.83
 
+## Fixed
+
+- Add a local copy of CSSBackgroundDrawable since it was removed from React Native
+
 ## [0.17.3] - 2026-01-03
 
 ## Fixed